requirements.txt
python-telegram-bot==20.7
anthropic==0.18.1
python-dotenv==1.0.0
requests==2.31.0
aiosqlite==0.19.0
beautifulsoup4==4.12.3
pytz==2024.1
